requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.25
scipy>=1.11
reportlab>=4.1
openpyxl>=3.1
matplotlib>=3.8
plotly>=5.18
scikit-learn>=1.4
statsmodels>=0.14
python-docx>=1.1
CoolProp>=6.4
nrel-pysam>=3.0
